Hoist the byte-parallel halves of InputPin/OutputPin into Pin (#27 S4)
The audited residuals that hoist cleanly: the pentagon draw body
(point-for-point identical in both classes), rotate/flip (now via the
orientation enum helpers, clearing both put lists), canRotate/canFlip,
the tristate load marker, currentValue/getCurrentValue, and
printValue/remove behind a pinKind() name hook. Pin's previous
remove(), which unregistered the name, is merged into the hoisted
subcircuit-guarded remove so the chain's combined behavior is
preserved. Kept per-leaf, as the audit prescribed: init geometry,
save, showInfo, setTriState, initSim, react, showCurrentValue.
