+++ src/jls/elem/Pin.java
@@ -4,7 +4,9 @@ import jls.*;
 
 import java.awt.*;
 import java.awt.event.*;
+import java.awt.geom.*;
 import java.io.*;
+import java.util.BitSet;
 
 import javax.swing.*;
 
@@ -204,7 +206,7 @@ public abstract class Pin extends LogicElement {
 	
 	/**
 	 * Save this element in a file.
-	 * 
+	 *
 	 * @param output The output writer.
 	 */
 	public void save(PrintWriter output) {
@@ -212,6 +214,237 @@ public abstract class Pin extends LogicElement {
 		super.save(output);
 		output.println("END");
 	} // end of save method
+
+	/**
+	 * The pin kind for user-facing messages ("Input" or "Output").
+	 */
+	protected abstract String pinKind();
+
+	/**
+	 * Load-time tri-state marker (" int tristate 1" in saved files).
+	 */
+	protected boolean loadTriState = false;
+
+	/**
+	 * Set an int instance variable value (during a load).
+	 *
+	 * @param name The instance variable name.
+	 * @param value The instance variable value.
+	 */
+	public void setValue(String name, int value) {
+
+		if (name.equals("tristate")) {
+			loadTriState = true;
+		} else {
+			super.setValue(name, value);
+		}
+	} // end of setValue method
+
+	/**
+	 * Draw this pin: the orientation-pointed pentagon, watched
+	 * background, centered name, and its single put (#27 S4 - the two
+	 * pin classes drew byte-identical shapes).
+	 *
+	 * @param g The graphics object to draw with.
+	 */
+	public void draw(Graphics g) {
+
+		// set up shape
+		int s = JLSInfo.spacing;
+		Polygon p = new Polygon();
+		if (orientation == JLSInfo.Orientation.RIGHT) {
+			p.addPoint(x, y);
+			p.addPoint(x + width - s, y);
+			p.addPoint(x + width, y + height / 2);
+			p.addPoint(x + width - s, y + height);
+			p.addPoint(x, y + height);
+		} else if (orientation == JLSInfo.Orientation.LEFT) {
+			p.addPoint(x + s, y);
+			p.addPoint(x, y + height / 2);
+			p.addPoint(x + s, y + height);
+			p.addPoint(x + width, y + height);
+			p.addPoint(x + width, y);
+		} else if (orientation == JLSInfo.Orientation.UP) {
+			p.addPoint(x + width / 2, y);
+			p.addPoint(x + width, y + s);
+			p.addPoint(x + width, y + height);
+			p.addPoint(x, y + height);
+			p.addPoint(x, y + s);
+		} else if (orientation == JLSInfo.Orientation.DOWN) {
+			p.addPoint(x, y);
+			p.addPoint(x + width, y);
+			p.addPoint(x + width, y + height - s);
+			p.addPoint(x + width / 2, y + height);
+			p.addPoint(x, y + height - s);
+		}
+		// draw watched background
+		if (watched) {
+			g.setColor(JLSInfo.watchColor);
+			g.fillPolygon(p);
+		}
+
+		// draw context
+		super.draw(g);
+
+		// draw box
+		g.setColor(Color.BLACK);
+		g.drawPolygon(p);
+
+		// draw name inside box
+		FontMetrics fm = g.getFontMetrics();
+		Rectangle2D t = fm.getStringBounds(name, g);
+		double tw = t.getWidth();
+		double th = t.getHeight();
+		int bx = x;
+		int by = y;
+		int bwidth = width;
+		int bheight = height;
+		if (orientation == JLSInfo.Orientation.LEFT) {
+			bx = x + s;
+			bwidth = width - s;
+		}
+		else if (orientation == JLSInfo.Orientation.RIGHT) {
+			bwidth = width - s;
+		}
+		else if (orientation == JLSInfo.Orientation.UP) {
+			by = y + s;
+			bheight = height - s;
+		}
+		else { // DOWN
+			bheight = height - s;
+		}
+		int dx = (int) ((bwidth - tw) / 2);
+		int dy = (int) ((bheight - th) / 2 + fm.getAscent());
+
+		g.drawString(name, bx + dx, by + dy);
+
+		// draw the put
+		for (Input in : inputs) {
+			in.draw(g);
+		}
+		for (Output out : outputs) {
+			out.draw(g);
+		}
+	} // end of draw method
+
+	/**
+	 * Print current value to stdout.
+	 *
+	 * @param qual Qualified subcircuit name.
+	 */
+	public void printValue(String qual) {
+
+		if (qual.equals("")) {
+			System.out.printf("%s Pin %s: %s\n", pinKind(), name,
+					BitSetUtils.toDisplay(currentValue, bits));
+		}
+		else {
+			System.out.printf("%s Pin %s.%s: %s\n", pinKind(), qual, name,
+					BitSetUtils.toDisplay(currentValue, bits));
+		}
+	} // end of printValue method
+
+	/**
+	 * Remove this element, but only if it is not part of a subcircuit.
+	 *
+	 * @param circ The circuit this element is in.
+	 */
+	public void remove(Circuit circ) {
+
+		if (circ.isImported()) {
+			JOptionPane.showMessageDialog(JLSInfo.frame,
+					"Can't remove " + pinKind().toLowerCase() + " pin "
+					+ name + " from a subcircuit");
+			return;
+		}
+		circ.removeName(name);
+		super.remove(circ);
+	} // end of remove method
+
+	/**
+	 * A pin can rotate or flip only when its put has no attachment.
+	 *
+	 * @return false if the put has a wire attached, true otherwise.
+	 */
+	public boolean canRotate() {
+
+		for (Input in : inputs) {
+			if (in.isAttached()) {
+				return false;
+			}
+		}
+		for (Output out : outputs) {
+			if (out.isAttached()) {
+				return false;
+			}
+		}
+		return true;
+	} // end of canRotate method
+
+	/**
+	 * A pin can rotate or flip only when its put has no attachment.
+	 *
+	 * @return false if the put has a wire attached, true otherwise.
+	 */
+	public boolean canFlip() {
+
+		return canRotate();
+	} // end of canFlip method
+
+	/**
+	 * Rotate this pin, rebuilding its geometry and put.
+	 *
+	 * @param direction The direction to rotate.
+	 * @param g The current graphics context for size recalculation.
+	 */
+	public void rotate(JLSInfo.Orientation direction, Graphics g) {
+
+		if (direction == JLSInfo.Orientation.LEFT) {
+			orientation = orientation.ccw();
+		}
+		else if (direction == JLSInfo.Orientation.RIGHT) {
+			orientation = orientation.cw();
+		}
+		inputs.clear();
+		outputs.clear();
+		width = 0;
+		height = 0;
+		init(g);
+	} // end of rotate method
+
+	/**
+	 * Flip this pin, rebuilding its geometry and put.
+	 *
+	 * @param g The current graphics context for size recalculation.
+	 */
+	public void flip(Graphics g) {
+
+		orientation = orientation.flipped();
+		inputs.clear();
+		outputs.clear();
+		width = 0;
+		height = 0;
+		init(g);
+	} // end of flip method
+
+	// -----------------------------------------------------------------
+	// Simulation state shared by both pins
+	// -----------------------------------------------------------------
+
+	protected BitSet currentValue = new BitSet();
+
+	/**
+	 * Get the current value.
+	 *
+	 * @return the current value.
+	 */
+	public BitSet getCurrentValue() {
+
+		if (currentValue == null)
+			return null;
+		else
+			return (BitSet) currentValue.clone();
+	} // end of getCurrentValue method
 	
 	/**
 	 * Display info about this element.
@@ -222,16 +455,6 @@ public abstract class Pin extends LogicElement {
 		
 		info.setText(bits + " bit input pin");
 	} // end of showInfo method
-	
-	/**
-	 * Remove this element.
-	 * Take name out of circuit's list of names used.
-	 */
-	public void remove(Circuit circ) {
-		
-		circ.removeName(name);
-		super.remove(circ);
-	} // end of remove method
 	
 	/**
 	 * Dialog box to set input pin characteristics.
